forumapi/requirement.txt
Django==2.2.6
django-cors-headers==3.2.1
django-redis==4.11.0
django-restframework==0.0.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
Pillow==7.0.0
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.3
redis==3.4.1
sqlparse==0.3.0
